backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
python-jose
tenacity
groq
litellm